Cleaner sanitizing and validation
