+++ app/Filament/Resources/CommandResource/Pages/EditCommand.php
@@ -3,22 +3,14 @@
 namespace App\Filament\Resources\CommandResource\Pages;
 
 use App\Filament\Resources\CommandResource;
-use App\Support\Sanitizer;
 use Filament\Actions;
 use Filament\Resources\Pages\EditRecord;
 
 class EditCommand extends EditRecord
 {
+	use CommandResource\Traits\PreparesForValidation;
+
     protected static string $resource = CommandResource::class;
-
-	protected function prepareForValidation($attributes)
-	{
-		$attributes['signature'] = Sanitizer::sanitize($attributes['signature']);
-
-		$attributes['description'] = Sanitizer::sanitize($attributes['description']);
-
-		return $attributes;
-	}
 
 	protected function getActions(): array
     {
