+++ app/Filament/Resources/CommandResource/Traits/PreparesForValidation.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace App\Filament\Resources\CommandResource\Traits;
+
+use App\Support\Sanitizer;
+
+trait PreparesForValidation
+{
+	protected function prepareForValidation($attributes)
+	{
+		$attributes['date'] = collect($attributes['data'])
+			->transform(static function ($item, $key) {
+				if ($item === null) return null;
+
+				return match ($key) {
+					'signature', 'description' => Sanitizer::sanitize($item)->toString(),
+					default => $item,
+				};
+			});
+
+		return $attributes;
+	}
+}
