+++ app/Filament/Resources/PostResource/Pages/EditPost.php
@@ -9,6 +9,8 @@ use Filament\Resources\Pages\EditRecord;
 
 class EditPost extends EditRecord
 {
+	use PostResource\Traits\PreparesForValidation;
+
     protected static string $resource = PostResource::class;
 
     protected function getHeaderActions(): array
@@ -17,16 +19,4 @@ class EditPost extends EditRecord
             Actions\DeleteAction::make(),
         ];
     }
-
-	protected function prepareForValidation($attributes)
-	{
-		$attributes['data'] = collect($attributes['data'])
-			->transform(static fn ($item, $key) => match($key) {
-				'title' => Sanitizer::sanitize($item)->toString(),
-				'slug' => Sanitizer::slug($item)->toString(),
-				default => $item,
-			});
-
-		return $attributes;
-	}
 }
